moving total event grade logic to a util file
